requirements.txt
Django>=5.0
djangorestframework
openai
psycopg2-binary
gunicorn
whitenoise
dj-database-url